requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
langgraph
langchain
python-dotenv
boto3>=1.34.0
httpx>=0.26.0
python-multipart>=0.0.9
pytest>=8.0.0
pypdf>=4.2.0
ipython
langchain-google-genai
requests-html
lxml_html_clean